Move `{hir::lowering -> hir}::is_range_literal`.
The function is never used inside lowering,
but only ever in external crates.

By moving it, we faciliate lowering as its own crate.
